feat: Add Slider | flexible compact redesign
Changed display and alignment to get  more centered and flexible container when scaling, also to create more space for the slider, so it won´t displace the chapter container while you aren´t in fullscreen.
Added two stacked slider, because i didn´t know a other way to increase the interactable area of the slider and change design while stay in html instyle, for faster and  more precise pb-values.